changed from dp to real

+++ Euler/basis.f90
@@ -10,15 +10,15 @@ subroutine basis_func_1d(phi,phi_xi,xi)
     use ader_dg
     implicit none
     ! Argument list
-    double precision, intent(in ) :: xi                    ! coordinate in [0,1] where to evaluate the basis
-    double precision, intent(out) :: phi(N+1), phi_xi(N+1) ! the basis and its derivative w.r.t. xi
+    real, intent(in ) :: xi                    ! coordinate in [0,1] where to evaluate the basis
+    real, intent(out) :: phi(N+1), phi_xi(N+1) ! the basis and its derivative w.r.t. xi
     ! Local variables
     integer            :: i,j,m
-    double precision   :: tmp
+    real   :: tmp
 
     ! Initialize variables
-    phi      = 1.0d0
-    phi_xi   = 0.0d0
+    phi      = 1.0
+    phi_xi   = 0.0
 
     ! Lagrange polynomial and its derivative
     do m = 1, N+1
@@ -28,7 +28,7 @@ subroutine basis_func_1d(phi,phi_xi,xi)
        end do
        do i = 1, N+1
           if(i .eq. m) cycle
-          tmp = 1.0d0
+          tmp = 1.0
           do j = 1, N+1
              if(j .eq. i) cycle
              if(j .eq. m) cycle
